crm_module/requirements.txt
# Requirements for CRM Module - Structured RAG System

# Environment variables management
python-dotenv

# Database and ORM
sqlalchemy
psycopg2-binary

# LangChain for SQL agent and OpenAI integration
langchain-openai
langchain-community

# Vector similarity search and embeddings
faiss-cpu
sentence-transformers

# Scientific computing and data manipulation
numpy

# Standard library dependencies (included for completeness)
# These are typically included with Python but listed for clarity
# os (built-in)
# datetime (built-in)
# json (built-in)
# pickle (built-in)